[fud] Add save_temps option to Xilinx synthesis (#851)
* First attempt at save_temps

* Fix abspath invocation

* Document save_temps

* Fix read_file output function

* Little docs fixes

* 2 minutes turns out to be what it takes to crash in synthesis. 5
  minutes is about what it takes to successfully produce an `xclbin`.
* The fud invocation should use `-o` instead of `--to` because the
  output file is binary. (fud will crash when trying to decode it to a
  printable string.)

* Fix close_and_get in SSH context

* Rename `keep` to `keep_tmpdir`

* Comment about save_temps

+++ fud/fud/stages/remote_context.py
@@ -120,10 +120,11 @@ class RemoteExecution:
 
         run_remote(client, tmpdir)
 
-    def _close(self, client, remote_tmpdir):
+    def _close(self, client, remote_tmpdir, keep_tmpdir=False):
         """Close the SSH connection to the server.
 
-        Also removes the remote temporary directory.
+        Also removes the remote temporary directory, unless the
+        `keep_tmpdir` flag is set.
         """
 
         @self.stage.step()
@@ -131,7 +132,8 @@ class RemoteExecution:
             """
             Remove created temporary files and close ssh connection.
             """
-            client.exec_command(f"rm -r {tmpdir}")
+            if not keep_tmpdir:
+                client.exec_command(f"rm -r {tmpdir}")
             client.close()
 
         finalize_ssh(client, remote_tmpdir)
@@ -159,7 +161,7 @@ class RemoteExecution:
         copy_back(client, remote_tmpdir, local_tmpdir)
         self._close(client, remote_tmpdir)
 
-    def close_and_get(self, client, remote_tmpdir, path):
+    def close_and_get(self, client, remote_tmpdir, path, keep_tmpdir=False):
         """Close the SSH connection and retrieve a single file.
 
         Produces the resulting downloaded file.
@@ -176,8 +178,8 @@ class RemoteExecution:
                 dest_path = tmpfile.name
             with self.SCPClient(client.get_transport()) as scp:
                 scp.get(src_path, dest_path)
-            return dest_path.open("rb")
+            return Path(dest_path)
 
         local_path = fetch_file(client, remote_tmpdir)
-        self._close(client, remote_tmpdir)
+        self._close(client, remote_tmpdir, keep_tmpdir=keep_tmpdir)
         return local_path
